[data, model] refactor: consolidate minimax_h3 code per review comments
- Move minimax_h3 preprocessor into shared dit/preprocess.py, drop duplicate module
- Remove minimax_h3-specific trainer, fold its logic into DiTTrainer
- Simplify data_transform selection: default to dit_online, keep dit_offline for offline_training
