+++ veomni/data/multimodal/dit/preprocess.py
@@ -1,6 +1,7 @@
 # dit preprocess should not be used for llm or mllms
+import os
+
 from ..preprocess import PREPROCESSOR_REGISTRY
-from . import minimax_h3_preprocess  # register minimax_h3 preprocessor  # noqa: F401
 
 
 @PREPROCESSOR_REGISTRY.register("Tom-and-Jerry-VideoGeneration-Dataset")
@@ -27,3 +28,25 @@ def qwen_image_preprocess(conversations, **kwargs):
     if image is None:
         raise ValueError("Qwen-Image data requires one of: image, image_bytes, image_path, target_image.")
     return prompt, {}, [image], []
+
+
+@PREPROCESSOR_REGISTRY.register("minimax_h3")
+def minimax_h3_preprocess(conversations, **kwargs):
+    data_dir = kwargs.get("data_dir", "")
+    prompt = conversations["prompt"]
+
+    # Video path
+    video_path = conversations.get("video", "")
+    if video_path and data_dir:
+        video_path = os.path.join(data_dir, video_path)
+
+    # Audio path (optional)
+    audio_path = conversations.get("input_audio", "")
+    if audio_path and data_dir:
+        audio_path = os.path.join(data_dir, audio_path)
+
+    audios = {"audio": audio_path} if audio_path else {}
+    videos = [video_path] if video_path else []
+
+    # FL2VA keyframes extracted from video frames in condition_model.get_condition()
+    return prompt, audios, [], videos
